gan/requirements.txt
tensorflow-gpu==1.15
matplotlib==3.3.4
seaborn==0.11.2
tqdm==4.64.1
wandb==0.13.3
